# docs/swagger.yaml
openapi: 3.0.0
info:
  title: FEDS Music API
  description: Swagger API documentation for user, musician, and admin endpoints authentication endpoint.
  version: 1.0.0
servers:
  - url: https://web-music-proj.onrender.com
paths:
  /user/add_user:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                email:
                  type: string
                name:
                  type: string
                task_id:
                  type: string
              required: [username, password, email, name, task_id]
      responses:
        '200':
          description: User added successfully
  /user/authenticate_user:
    post:
      summary: Authenticate user login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required: [username, password]
      responses:
        '200':
          description: User authenticated successfully
...